refactor: isolate local AI providers from upstream code
Move Voyage/Cohere/Ollama implementations into dedicated modules so
upstream files carry only short dispatch hooks (~19 lines total):

- embed/voyage.rs: VoyageEmbedder (was inline in embed/mod.rs)
- rerank.rs: maybe_rerank() single-call hook for Cohere second-stage
  rerank (was a 36-line block in mcp tools/search.rs)
- review/ollama.rs: Ollama review backend (was inline in review/llm.rs)

Behavior unchanged: providers activate only via VOYAGE_API_KEY /
COHERE_API_KEY / OLLAMA_API_KEY; without keys everything is fully
upstream-compatible and offline. Eases future rebases onto upstream.

Amp-Thread-ID: https://ampcode.com/threads/T-019f9980-7174-7219-8f5c-3f70ad45231a
