+++ crates/infigraph-core/src/embed/mod.rs
@@ -6,6 +6,9 @@ use anyhow::{Context, Result};
 
 use crate::model::Symbol;
 
+mod voyage; // local-dev add-on: remote Voyage AI embedder
+pub use voyage::VoyageEmbedder;
+
 struct CachedEmbeddings {
     path: PathBuf,
     modified: std::time::SystemTime,
@@ -304,107 +307,13 @@ impl EmbedProvider for Model2VecEmbedder {
     }
 }
 
-/// Remote embedder using the Voyage AI embeddings API.
-///
-/// Activated when `VOYAGE_API_KEY` is set. Uses `voyage-code-3` by default
-/// (override with `INFIGRAPH_VOYAGE_MODEL`). Requests 256-dimensional output
-/// (Matryoshka truncation) so vectors stay compatible with the local
-/// embeddings.bin format, the HNSW index, and the Postgres `vector(256)` column.
-pub struct VoyageEmbedder {
-    api_key: String,
-    model: String,
-    dim: usize,
-}
-
-impl VoyageEmbedder {
-    /// Dimension pinned to 256 to match the pgvector schema (`vector(256)`).
-    pub const OUTPUT_DIMENSION: usize = 256;
-    const ENDPOINT: &'static str = "https://api.voyageai.com/v1/embeddings";
-    /// Voyage allows up to 1000 inputs per request; stay well under token limits.
-    const BATCH_SIZE: usize = 128;
-
-    /// Build from `VOYAGE_API_KEY`; returns None when the key is not set.
-    pub fn from_env() -> Option<Self> {
-        let api_key = std::env::var("VOYAGE_API_KEY")
-            .ok()
-            .filter(|k| !k.is_empty())?;
-        let model =
-            std::env::var("INFIGRAPH_VOYAGE_MODEL").unwrap_or_else(|_| "voyage-code-3".to_string());
-        Some(Self {
-            api_key,
-            model,
-            dim: Self::OUTPUT_DIMENSION,
-        })
-    }
-
-    fn request_batch(&self, texts: &[&str]) -> Result<Vec<Vec<f32>>> {
-        let body = serde_json::json!({
-            "input": texts,
-            "model": self.model,
-            "output_dimension": self.dim,
-        });
-        let resp = ureq::post(Self::ENDPOINT)
-            .set("Authorization", &format!("Bearer {}", self.api_key))
-            .set("Content-Type", "application/json")
-            .send_json(body)
-            .map_err(|e| anyhow::anyhow!("voyage embeddings request failed: {e}"))?;
-        let json: serde_json::Value = resp
-            .into_json()
-            .context("voyage embeddings: invalid JSON response")?;
-        let data = json["data"]
-            .as_array()
-            .ok_or_else(|| anyhow::anyhow!("voyage embeddings: missing 'data' in response"))?;
-        anyhow::ensure!(
-            data.len() == texts.len(),
-            "voyage embeddings: expected {} vectors, got {}",
-            texts.len(),
-            data.len()
-        );
-        // Order by 'index' to be robust against out-of-order responses.
-        let mut out = vec![Vec::new(); texts.len()];
-        for item in data {
-            let idx = item["index"].as_u64().unwrap_or(0) as usize;
-            let emb: Vec<f32> = item["embedding"]
-                .as_array()
-                .ok_or_else(|| anyhow::anyhow!("voyage embeddings: missing 'embedding'"))?
-                .iter()
-                .filter_map(|v| v.as_f64().map(|f| f as f32))
-                .collect();
-            anyhow::ensure!(
-                emb.len() == self.dim,
-                "voyage embeddings: expected dim {}, got {}",
-                self.dim,
-                emb.len()
-            );
-            anyhow::ensure!(idx < out.len(), "voyage embeddings: index out of range");
-            out[idx] = emb;
-        }
-        Ok(out)
-    }
-}
-
-impl EmbedProvider for VoyageEmbedder {
-    fn dimension(&self) -> usize {
-        self.dim
-    }
-
-    fn embed_batch(&self, texts: &[&str]) -> Result<Vec<Vec<f32>>> {
-        let mut all = Vec::with_capacity(texts.len());
-        for chunk in texts.chunks(Self::BATCH_SIZE) {
-            all.extend(self.request_batch(chunk)?);
-        }
-        Ok(all)
-    }
-}
-
 static CODE_EMBEDDER: OnceLock<Arc<dyn EmbedProvider>> = OnceLock::new();
 static DOC_EMBEDDER: OnceLock<Arc<dyn EmbedProvider>> = OnceLock::new();
 
-/// Factory: prefer Voyage AI when `VOYAGE_API_KEY` is set, then Model2Vec,
-/// otherwise fall back to TrigramEmbedder.
+/// Factory: select Model2Vec if available, otherwise fall back to TrigramEmbedder.
 pub fn init_embedder() -> Arc<dyn EmbedProvider> {
     if let Some(v) = VoyageEmbedder::from_env() {
-        return Arc::new(v);
+        return Arc::new(v); // local-dev hook: VOYAGE_API_KEY takes precedence
     }
     match Model2VecEmbedder::new() {
         Ok(m) => Arc::new(m),
@@ -425,11 +334,10 @@ pub fn doc_embedder() -> Arc<dyn EmbedProvider> {
     Arc::clone(DOC_EMBEDDER.get_or_init(init_embedder))
 }
 
-/// Create the best available embedder: Voyage AI when `VOYAGE_API_KEY` is set,
-/// then Model2Vec if possible, fallback to trigram.
+/// Create the best available embedder: Model2Vec if possible, fallback to trigram.
 pub fn best_embedder() -> Box<dyn EmbedProvider> {
     if let Some(v) = VoyageEmbedder::from_env() {
-        return Box::new(v);
+        return Box::new(v); // local-dev hook: VOYAGE_API_KEY takes precedence
     }
     match Model2VecEmbedder::new() {
         Ok(m) => Box::new(m),
