+++ crates/infigraph-core/src/rerank.rs
@@ -1,11 +1,14 @@
 //! Optional remote reranking via the Cohere Rerank API.
 //!
-//! Activated when `COHERE_API_KEY` is set. Uses `rerank-v3.5` by default
-//! (override with `INFIGRAPH_COHERE_RERANK_MODEL`). Callers should treat this
-//! as a best-effort second-stage reranker: on any API failure, fall back to
-//! the local score ordering.
+//! Local-development add-on: everything Cohere-specific lives in this file so
+//! callers only need a single `maybe_rerank(...)` hook line. Activated when
+//! `COHERE_API_KEY` is set. Uses `rerank-v3.5` by default (override with
+//! `INFIGRAPH_COHERE_RERANK_MODEL`). Best-effort second-stage reranker: on any
+//! API failure, the local score ordering is kept.
 
 use anyhow::{Context, Result};
+
+use crate::search::SearchResult;
 
 const ENDPOINT: &str = "https://api.cohere.com/v2/rerank";
 /// Cohere recommends at most 1000 documents per request; we cap far lower
@@ -21,6 +24,56 @@ pub fn cohere_enabled() -> bool {
 
 fn model() -> String {
     std::env::var("INFIGRAPH_COHERE_RERANK_MODEL").unwrap_or_else(|_| "rerank-v3.5".to_string())
+}
+
+/// Single-call hook for search pipelines: when `COHERE_API_KEY` is set, rerank
+/// the top candidates of `results` in place (scores replaced with Cohere
+/// relevance in `[0, 1]`). `docs` maps symbol IDs to embedding text; symbols
+/// missing from it fall back to a `kind name in file` description. Best-effort:
+/// on any API failure the local ordering is kept and a warning is printed.
+pub fn maybe_rerank(
+    query: &str,
+    results: &mut Vec<SearchResult>,
+    docs: &[(String, String)],
+    limit: usize,
+) {
+    if !cohere_enabled() || results.is_empty() {
+        return;
+    }
+    let candidate_count = results
+        .len()
+        .min(limit.saturating_mul(2).max(limit))
+        .min(100);
+    let doc_text: std::collections::HashMap<&str, &str> = docs
+        .iter()
+        .map(|(id, text)| (id.as_str(), text.as_str()))
+        .collect();
+    let texts: Vec<String> = results[..candidate_count]
+        .iter()
+        .map(|r| {
+            doc_text
+                .get(r.symbol_id.as_str())
+                .map(|t| (*t).to_string())
+                .unwrap_or_else(|| format!("{} {} in {}", r.kind, r.name, r.file))
+        })
+        .collect();
+    match cohere_rerank(query, &texts, candidate_count) {
+        Ok(ranked) => {
+            let mut reordered: Vec<SearchResult> = ranked
+                .iter()
+                .filter_map(|(idx, score)| {
+                    results.get(*idx).map(|r| {
+                        let mut r = r.clone();
+                        r.score = *score;
+                        r
+                    })
+                })
+                .collect();
+            reordered.extend(results[candidate_count..].iter().cloned());
+            *results = reordered;
+        }
+        Err(e) => eprintln!("warning: cohere rerank failed, using local ranking: {e}"),
+    }
 }
 
 /// Rerank `documents` against `query` with the Cohere Rerank API.
